fix: make snapshot boot work in production
Snapshot fast-boot was inert in prod. The workspace wrote snapshots to
<root>/.outl/snapshots but storage read them back from <root>/snapshots
(it derived the dir from ops_dir.parent(), and prod keeps the op log at
<root>/ops). Writer and reader never met, so every boot fell back to a
full replay. The tests passed only because they use <root>/.outl/ops,
which makes the two dirs coincide.

Snapshot I/O now belongs to Workspace, keyed off the workspace root, and
is off the Storage trait, so there is a single path derivation and the
two-owners divergence can't come back. Fixing only the path would arm a
silent data-loss bug: the replay cutoff was one global HLC, so a low-HLC
op from a lagging peer delivered after the snapshot sat below it and
vanished from the tree though it was durably in storage. The cutoff is
now a per-actor vector clock, so snapshot boot stays observationally
identical to a full replay.

The TUI, desktop, and mobile enable the snapshot policy on open and flush
a final snapshot on shutdown, so the CLI boots from one instead of
replaying the whole op log.

Also splits jsonl.rs (1189 lines) into jsonl/{mod,read,append,tests},
a pure restructure with no logic change.

Fixes #156
Fixes #128
Fixes #109
Fixes #161
